+++ tests/corpus/io/test_kaldi.py
@@ -1,8 +1,12 @@
 import os
 
+import numpy as np
+
 from audiomate import corpus
+from audiomate import tracks
 from audiomate import issuers
 from audiomate.corpus import io
+from audiomate.utils import textfile
 
 import pytest
 
@@ -132,3 +136,99 @@ class TestKaldiWriter:
         assert 'utt2spk' in os.listdir(path)
         assert 'spk2gender' in os.listdir(path)
         assert 'wav.scp' in os.listdir(path)
+
+    def test_write_wav_scp(self, writer, tmpdir):
+        ds = resources.create_dataset()
+        path = tmpdir.strpath
+        writer.save(ds, path)
+
+        content = textfile.read_separated_lines(
+            os.path.join(path, 'wav.scp'),
+            separator=' ',
+            max_columns=2
+        )
+
+        wav_base = resources.get_resource_path(['wav_files'])
+        wav_base = os.path.abspath(wav_base)
+
+        assert content[0][0] == 'wav-1'
+        assert content[0][1] == os.path.join(wav_base, 'wav_1.wav')
+        assert content[1][0] == 'wav_2'
+        assert content[1][1] == os.path.join(wav_base, 'wav_2.wav')
+        assert content[2][0] == 'wav_3'
+        assert content[2][1] == os.path.join(wav_base, 'wav_3.wav')
+        assert content[3][0] == 'wav_4'
+        assert content[3][1] == os.path.join(wav_base, 'wav_4.wav')
+
+    def test_write_segments(self, writer, tmpdir):
+        ds = resources.create_dataset()
+        path = tmpdir.strpath
+        writer.save(ds, path)
+
+        content = textfile.read_separated_lines(
+            os.path.join(path, 'segments'),
+            separator=' ',
+            max_columns=4
+        )
+
+        assert content[0][0] == 'utt-1'
+        assert content[0][1] == 'wav-1'
+        assert float(content[0][2]) == pytest.approx(0)
+        assert float(content[0][3]) == pytest.approx(2.5951875)
+
+        assert content[1][0] == 'utt-2'
+        assert content[1][1] == 'wav_2'
+        assert float(content[1][2]) == pytest.approx(0)
+        assert float(content[1][3]) == pytest.approx(2.5951875)
+
+        assert content[2][0] == 'utt-3'
+        assert content[2][1] == 'wav_3'
+        assert float(content[2][2]) == pytest.approx(0)
+        assert float(content[2][3]) == pytest.approx(1.5)
+
+        assert content[3][0] == 'utt-4'
+        assert content[3][1] == 'wav_3'
+        assert float(content[3][2]) == pytest.approx(1.5)
+        assert float(content[3][3]) == pytest.approx(2.5)
+
+        assert content[4][0] == 'utt-5'
+        assert content[4][1] == 'wav_4'
+        assert float(content[4][2]) == pytest.approx(0)
+        assert float(content[4][3]) == pytest.approx(2.5951875)
+
+    def test_exports_wavs_from_container_tracks(self, writer, tmpdir):
+        path = tmpdir.strpath
+        container_ds_path = os.path.join(path, 'container_ds')
+        out_path = os.path.join(path, 'export')
+
+        ds = resources.create_dataset()
+        ds.relocate_audio_to_single_container(container_ds_path)
+
+        writer.save(ds, out_path)
+
+        print(os.listdir(out_path))
+
+        track_path = os.path.join(out_path, 'audio', 'wav-1.wav')
+        track = tracks.FileTrack(None, track_path)
+        assert os.path.isfile(track_path)
+        assert track.duration == pytest.approx(2.5951875)
+        assert np.allclose(
+            track.read_samples(),
+            ds.tracks['wav-1'].read_samples(),
+            atol=1e-05
+        )
+
+        track_path = os.path.join(out_path, 'audio', 'wav_2.wav')
+        track = tracks.FileTrack(None, track_path)
+        assert os.path.isfile(track_path)
+        assert track.duration == pytest.approx(2.5951875)
+
+        track_path = os.path.join(out_path, 'audio', 'wav_3.wav')
+        track = tracks.FileTrack(None, track_path)
+        assert os.path.isfile(track_path)
+        assert track.duration == pytest.approx(2.5951875)
+
+        track_path = os.path.join(out_path, 'audio', 'wav_4.wav')
+        track = tracks.FileTrack(None, track_path)
+        assert os.path.isfile(track_path)
+        assert track.duration == pytest.approx(2.5951875)